Conflicts:
	Controller/MarketplacesController.php
	companyCodeFiles/p2pCompany.class.php
	companyCodeFiles/zank.php

+++ companyCodeFiles/p2pCompany.class.php
@@ -624,7 +624,6 @@ class p2pCompany {
     function doCompanyLoginMultiCurl(array $loginCredentials) {
         echo 'credentials: ' . print_r($loginCredentials);
         $url = array_shift($this->urlSequence);
-        echo 'login:' . $url;
         $this->errorInfo = $url;
         if (!empty($this->testConfig['active']) == true) {  // test system active, so read input from prepared files
             if (!empty($this->testConfig['siteReadings'])) {
@@ -680,7 +679,7 @@ class p2pCompany {
         
         $request->_page = json_encode($info);
         // Add the url to the queue
-        $this->classContainer->addRequetsToQueueCurls($request);
+        $this->classContainer->addRequestToQueueCurls($request);
     }
 
     /**
@@ -759,7 +758,7 @@ class p2pCompany {
                 ->set(CURLOPT_COOKIEFILE, $this->cookiesDir . '/' . $this->cookies_name)
                 ->set(CURLOPT_COOKIEJAR, $this->cookiesDir . '/' . $this->cookies_name);
 
-        $this->classContainer->addRequetsToQueueCurls($request);
+        $this->classContainer->addRequestToQueueCurls($request);
     }
 
     /**
@@ -803,7 +802,6 @@ class p2pCompany {
 
         // check if extra headers have to be added to the http message  
         if (!empty($this->headers)) {
-            echo "EXTRA HEADERS TO BE ADDED<br>";
             $request->getOptions()
                     //->set(CURLOPT_HEADER, true) Esto fue una prueba, no funciona, quitar
                     ->set(CURLOPT_HTTPHEADER, $this->headers);
@@ -841,7 +839,7 @@ class p2pCompany {
                 ->set(CURLOPT_COOKIEFILE, $this->cookiesDir . '/' . $this->cookies_name) // important
                 ->set(CURLOPT_COOKIEJAR, $this->cookiesDir . '/' . $this->cookies_name); // Important
         //Add the request to the queue in the classContainer controller
-        $this->classContainer->addRequetsToQueueCurls($request);
+        $this->classContainer->addRequestToQueueCurls($request);
 
         if ($this->config['appDebug'] == true) {
             echo "VISITED COMPANY URL = $url <br>";
@@ -1380,7 +1378,7 @@ class p2pCompany {
      * Delete the cookies file generated for the request
      */
     public function deleteCookiesFile() {
-        if (file_exists($this->cookiesDir . '/' . $this->cookies_name)) {
+        if ($this->cookies_name != "cookies.txt" && file_exists($this->cookiesDir . '/' . $this->cookies_name)) {
             unlink($this->cookiesDir . '/' . $this->cookies_name);
         }
     }
@@ -1844,7 +1842,7 @@ class p2pCompany {
                 ->set(CURLOPT_COOKIEFILE, $this->cookiesDir . '/' . $this->cookies_name) // important
                 ->set(CURLOPT_COOKIEJAR, $this->cookiesDir . '/' . $this->cookies_name); // Important
         //Add the request to the queue in the classContainer controller
-        $this->classContainer->addRequetsToQueueCurls($request);
+        $this->classContainer->addRequestToQueueCurls($request);
         
         
         /*
@@ -1891,11 +1889,11 @@ class p2pCompany {
                       echo $nameAttrNode2 . '=>' . $valueAttrNode2 . '<br>'; */
 
                     if ($nameAttrNode1 != $nameAttrNode2) {
-                        //echo 'Node attr name error';
+                        echo 'Node attr name error';
                         $this->same_structure = false;
                     }
                     if ($valueAttrNode1 != $valueAttrNode2) {
-                        //echo 'Node attr value error';
+                        echo 'Node attr value error';
                         $this->sameStructure = false;
                     }
                     if ($this->sameStructure) {
@@ -1916,12 +1914,12 @@ class p2pCompany {
                 $this->sameStructure = false;
             }
         } else if ($node1->hasAttributes() && !$node2->hasAttributes()) {
-            //echo $node1->tagName . ' / ' . $node2->tagName . '<br>';
-            //echo 'Node2 has attr error';
+            echo $node1->tagName . ' / ' . $node2->tagName . '<br>';
+            echo 'Node2 has attr error';
             $this->sameStructure = false;
         } else if (!$node1->hasAttributes() && $node2->hasAttributes()) {
-            //echo $node1->tagName . ' / ' . $node2->tagName . '<br>';
-            //echo 'Node1 has attr error';
+            echo $node1->tagName . ' / ' . $node2->tagName . '<br>';
+            echo 'Node1 has attr error';
             $this->sameStructure = false;
         }
         if ($this->sameStructure && !$repeatedStructureFound) {
@@ -1957,13 +1955,13 @@ class p2pCompany {
 
                     $this->verifyDomStructure($childrenNode1[$i], $childrenNode2[$i], $uniquesElement, $limit);
                 }
-            } else if (!$node1->hasChildNodes() && $node2->hasChildNodes()) {
-                //echo 'Node has attr error 2';
+            } /*else if (!$node1->hasChildNodes() && $node2->hasChildNodes()) {
+                echo 'Node has attr error 2';
                 $this->sameStructure = false;
             } else if ($node1->hasChildNodes() && !$node2->hasChildNodes()) {
-                //echo 'Node has attr error 2';
+                echo 'Node has attr error 2';
                 $this->sameStructure = false;
-            }
+            }*/
         }
         return $this->sameStructure;
     }
@@ -2225,6 +2223,26 @@ class p2pCompany {
         }
         return [$structureRevision, $break, $type];
     }
+    
+    
+    /**Search in the pfp marketplace the winvestify marketplace loan id. If we find it we can delete from the array.
+     * The array will contain the deleted/hidden invesment that we cant update from the pfp marketplace.
+     * @param array $loanReferenceList loan reference id list that we have in our marketplace
+     * @param array $investment single investment that we compare
+     */
+    public function marketplaceLoanIdWinvestifyPfpComparation($loanReferenceList,$investment){  
+        print_r($investment);
+        print_r($loanReferenceList);
+         foreach($loanReferenceList as $key => $winvestifyMarketplaceLoanId){
+            if($winvestifyMarketplaceLoanId == $investment['marketplace_loanReference']){
+                echo 'Loan finded, deleting from array' . HTML_ENDOFLINE;
+                unset($loanReferenceList[$key]); 
+            }
+        }
+        
+        return $loanReferenceList;
+    }
+    
 
 }
